school admin : dashboard : updated attendence pie chart respone

+++ src/main/java/com/techjagannath/digitalidentification/repository/customrepositories/SchoolAdminCustomRepository.java
@@ -1,5 +1,5 @@
 package com.techjagannath.digitalidentification.repository.customrepositories;
 
 public interface SchoolAdminCustomRepository {
-    Long retrieveAttendancePieChartData(Long schoolId);
+    Integer retrieveAttendancePieChartData(Long schoolId);
 }
